pyproject.toml
[project]
name = "tivim8"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "pycryptodome>=3.23.0",
    "uvicorn>=0.34.3",
]
